clang 6 unintentional conversions fixed

+++ include/config_parser/config_parser.h
@@ -24,7 +24,7 @@ public:
 
   int pause() const;
 
-  size_t publishers() const;
+  int publishers() const;
 
   std::string host() const;
